[refactor] devlop 브랜치와의 결합을 위한 경로수정
