+++ src/main/java/com/haejwo/tripcometrue/domain/triprecord/repository/triprecord/TripRecordCustomRepository.java
@@ -1,7 +1,7 @@
 package com.haejwo.tripcometrue.domain.triprecord.repository.triprecord;
 
 import com.haejwo.tripcometrue.domain.triprecord.dto.request.ModelAttribute.TripRecordListRequestAttribute;
-import com.haejwo.tripcometrue.domain.triprecord.dto.response.TripRecordListResponseDto;
+import com.haejwo.tripcometrue.domain.triprecord.dto.response.triprecord.TripRecordListResponseDto;
 import com.haejwo.tripcometrue.domain.triprecord.entity.TripRecord;
 import java.util.List;
 import org.springframework.data.domain.Pageable;
